problemset-175888/upload_center/requirements.txt
Django==4.1
djangorestframework==3.13.1